Navigate to xpromo links in the click-handler's call-stack.
Navigating to xpromo links has different behaviors per mobile
operating system that depends on (a) when the navigation starts,
(b) is this url a local/dev url, or a production url, and
(c) iOS or Android? There a few broken intermediate stages where
deep-linking for devices that have the Reddit app don't work,
or in dev on Android you see a broken intent url. All of these
are solved by navigating to the xpromo install links in the
click-handler's call-stack, before calling await.